added confirm for delete and icon

+++ src/features/Tasks/TaskDetails/TaskDetails.jsx
@@ -41,6 +41,10 @@ function TaskDetails({ task, onClose }) {
                     onBlur={(e) => handleBlurUpdate('description', e)}
                 />
                 <MdDeleteOutline className={"delete-icon"} onClick={() => {
+                    e.stopPropagation();
+                    if (!confirm("Are you sure you want to delete this task?")) {
+                        return;
+                    }
                     DeleteTask(task.id);
                 }} />
